Issue 52839: LKSM: Include the ability to use lineage language in aliquot naming patterns
